网关微服务平台初始化-优化后端代码用以支撑前端底座框架变更

+++ gateway/gateway-admin-server/src/main/java/cn/cleanarch/gw/gateway/admin/system/controller/SysDeptController.java
@@ -70,7 +70,7 @@ public class SysDeptController {
      */
     @DeleteMapping("/{id}")
     //@PreAuthorize("@pms.hasPermission('sys_dept_del')")
-    public R<Boolean> removeById(@PathVariable Integer id) {
+    public R<Boolean> removeById(@PathVariable Long id) {
         return R.success(sysDeptService.removeDeptById(id));
     }
