+++ gateway/gateway-admin-server/src/main/java/cn/cleanarch/gw/gateway/admin/system/service/SysDeptService.java
@@ -37,7 +37,7 @@ public interface SysDeptService extends IService<SysDept> {
      * @param id 部门 ID
      * @return 成功、失败
      */
-    Boolean removeDeptById(Integer id);
+    Boolean removeDeptById(Long id);
 
     /**
      * 更新部门
