+++ gateway/gateway-admin-server/src/main/java/cn/cleanarch/gw/gateway/admin/system/service/impl/SysDeptRelationServiceImpl.java
@@ -61,7 +61,7 @@ public class SysDeptRelationServiceImpl extends ServiceImpl<SysDeptRelationMappe
      * @param id
      */
     @Override
-    public void deleteAllDeptRelation(Integer id) {
+    public void deleteAllDeptRelation(Long id) {
         baseMapper.deleteDeptRelationsById(id);
     }
